1. springboot版本更换为2.2.13.RELEASE 2. mybatisplus更新到3.5.10.1 -- 由于mybatisplus版本升级到3.5.10.1版本，且使用了分页插件PaginationInnerInterceptor，所以按照mybatisplus介绍，关于mybatisplus部分使用了Maven bom方式处理。 使用 maven bom 管理依赖，减少版本号的冲突。因为 jsqlparser 5.0+ 版本不再支持 jdk8 针对这个问题解耦 jsqlparser 依赖。 3. 处理各种常识性以及安全问题

+++ src/main/java/com/mysiteforme/admin/service/MenuService.java
@@ -1,10 +1,8 @@
 package com.mysiteforme.admin.service;
 
-import com.baomidou.mybatisplus.service.IService;
+import com.baomidou.mybatisplus.extension.service.IService;
 import com.mysiteforme.admin.entity.Menu;
-import com.mysiteforme.admin.entity.User;
 import com.mysiteforme.admin.entity.VO.ShowMenu;
-import com.mysiteforme.admin.entity.VO.TreeMenu;
 import com.mysiteforme.admin.entity.VO.ZtreeVO;
 
 import java.util.List;
@@ -28,8 +26,8 @@ public interface MenuService extends IService<Menu> {
 
     void saveOrUpdateMenu(Menu menu);
 
-    int getCountByPermission(String permission);
+    long getCountByPermission(String permission);
 
-    int getCountByName(String name);
+    long getCountByName(String name);
 
 }
